requirements.txt
absl-py==1.2.0
dm-haiku==0.0.9
ml_collections==0.1.1
numpy>=1.21.0
jraph>=0.0.5.dev0
scikit-learn>=0.24.0
matplotlib>=3.5.0
dm-tree==0.1.7
plotly==5.11.0
typing_extensions==4.4.0
